requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
langchain
langchain-community
langchain-core
langchain-ollama
nltk==3.8.1
pandas==2.2.2
numpy==1.26.4
seaborn==0.12.2
langid==1.1.6
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.0.1
openai==1.3.7
transformers==4.35.2
scikit-learn==1.3.2
textblob==0.17.1
vaderSentiment==3.3.2
spacy==3.7.2
python-multipart==0.0.6
aiofiles==23.2.1
youtube-transcript-api==0.6.1
pydantic==2.4.2 